refactor: . Move QuickSearchOptionsModal to ui/ directory
