+++ src/ui/QuickSearchOptionsModal.ts
